python/requirements.txt
boto3
pycrypto